+++ tests/trends/test_trends.py
@@ -4,7 +4,8 @@ from conftest import APPROX_REL, LOGLEVEL
 from aqua.diagnostics.trends import Trends
 
 loglevel = LOGLEVEL
-approx_rel = APPROX_REL*10
+approx_rel = APPROX_REL * 10
+
 
 @pytest.mark.diagnostics
 def test_trends():
@@ -16,10 +17,12 @@ def test_trends():
 
     trend.run(
         # dim_mean="lat",
-              var=['thetao', 'so'],
-              region='go'
-              )
+        var=["thetao", "so"],
+        region="go",
+    )
     print(trend.trend_coef)
     assert trend is not None, "trend instance should not be None"
-    assert trend.trend_coef["thetao"].isel(level=1).mean('lat').mean('lon').values == pytest.approx(-0.06603967,rel=approx_rel)
-    assert trend.trend_coef["so"].isel(level=1).mean('lat').mean('lon').values == pytest.approx(0.02622599,rel=approx_rel)
+    assert trend.trend_coef["thetao"].isel(level=1).mean("lat").mean("lon").values == pytest.approx(
+        -0.06603967, rel=approx_rel
+    )
+    assert trend.trend_coef["so"].isel(level=1).mean("lat").mean("lon").values == pytest.approx(0.02622599, rel=approx_rel)